Refactor protect structure. Take TopMenu folders 1 level up
